feat: Adds handleTransaction to FactoryConfig.
